Rewrite PG url
