+++ votey/utils.py
@@ -1,14 +1,32 @@
+from dataclasses import dataclass
 from typing import Any
 from typing import Dict
 from typing import Iterable
 from typing import List
+from typing import Optional
 from typing import TypeVar
+from urllib.parse import urlparse
 
 T = TypeVar("T")
 
 JSON = Dict[str, str]
 ListJSON = Dict[str, List[JSON]]
 AnyJSON = Dict[str, Any]
+
+
+@dataclass
+class OptionData:
+    text: str
+    emoji: Optional[str]
+
+
+@dataclass
+class Command:
+    question: str
+    options: List[OptionData]
+    anonymous: bool
+    secret: bool
+    vote_emoji: Optional[str]
 
 
 def batch(lst: List[T], n: int = 1) -> Iterable[List[T]]:
@@ -24,3 +42,8 @@ def get_footer(user_id: str, anonymous: bool, secret: bool) -> str:
         return f"Anonymous poll created by <@{user_id}> with /votey"
 
     return f"Poll created by <@{user_id}> with /votey"
+
+
+def rewrite_pg_url(database_url: str) -> str:
+    url = urlparse(database_url)
+    return url._replace(scheme="postgresql+psycopg2").geturl()
